# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Talent Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Registration accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegisterResponse" }
        "400": { description: Invalid input }
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401": { description: Invalid credentials }
        "403": { description: Account not active }
...
  /auth/setup-password:
    post:
      operationId: setupPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupPasswordRequest" }
      responses:
        "200":
          description: Password set, user activated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { description: Invalid token or password }
  /auth/forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request a password reset link (always returns ok=true)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForgotPasswordRequest" }
      responses:
        "200":
          description: Request acknowledged (does not reveal whether email is registered)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400": { description: Invalid input }
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400": { description: Invalid input }
        "401": { description: "Not authenticated or current password incorrect" }
  /auth/setup-token/{token}:
    get:
      operationId: getSetupTokenInfo
      tags: [auth]
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Token info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SetupTokenInfo" }
        "404": { description: Invalid or expired token }
...
components:
  schemas:
...
    AuthUser:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        fullName: { type: string }
        role:
          type: string
          enum: [candidate, employer, institution, admin]
        status:
          type: string
          enum: [pending, active, rejected, invited]
        orgRole: { type: string, nullable: true }
        candidateId: { type: integer, nullable: true }
        employerId: { type: integer, nullable: true }
        institutionId: { type: integer, nullable: true }
        permissions:
          type: array
          items: { type: string }
          description: Effective admin permission keys; empty for non-admins.
      required:
        [id, email, fullName, role, status, orgRole, candidateId, employerId, institutionId, permissions]
    AuthSession:
      type: object
      properties:
        user:
          oneOf:
            - { $ref: "#/components/schemas/AuthUser" }
            - { type: "null" }
      required: [user]
    RegisterRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
        fullName: { type: string }
        role:
          type: string
          enum: [candidate, employer, institution]
        submittedData:
          type: object
          additionalProperties: true
      required: [email, password, fullName, role]
    RegisterResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    SetupPasswordRequest:
      type: object
      properties:
        token: { type: string }
        password: { type: string }
      required: [token, password]
    ForgotPasswordRequest:
      type: object
      properties:
        email: { type: string }
      required: [email]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
      required: [currentPassword, newPassword]
    SetupTokenInfo:
      type: object
      properties:
        email: { type: string }
        fullName: { type: string }
        role: { type: string }
      required: [email, fullName, role]
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...