# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Talent Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from "./authUserRole";
import type { AuthUserStatus } from "./authUserStatus";

export interface AuthUser {
  id: number;
  email: string;
  fullName: string;
  role: AuthUserRole;
  status: AuthUserStatus;
  orgRole: string | null;
  candidateId: number | null;
  employerId: number | null;
  institutionId: number | null;
  /** Effective admin permission keys; empty for non-admins. */
  permissions: string[];
}